% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_scoreid.R
\name{assign_pressure}
\alias{assign_pressure}
\title{Add Pressure Score to Point-level Data}
\usage{
assign_pressure(data)
}
\arguments{
\item{data}{Dataframe of points and events}
}
\description{
Add Pressure Score to Point-level Data
}
