% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_connection.R
\name{make_connection}
\alias{make_connection}
\title{Establish a Databse Connection}
\usage{
make_connection(driver = RMariaDB::MariaDB(), group = "tennis")
}
\value{
An S4 object that inherits from DBIConnection
}
\description{
Establish a Databse Connection
}
