% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchids.R
\name{matchids}
\alias{matchids}
\title{Collect URLs and Match Meta Data}
\usage{
matchids(the_tour, id, the_year)
}
\arguments{
\item{the_tour}{Name of tour level as in \code{tournaments}}

\item{id}{Event id as in \code{tournaments}}

\item{the_year}{Year of the event}
}
\description{
This function obtains urls and match-level info for all rounds at the specified event and tour
}
