% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp.R
\name{pbp}
\alias{pbp}
\title{Collect Point-by-Point Data}
\usage{
pbp(id, correct_point_sequences = TRUE)
}
\arguments{
\item{id}{Match id}
}
\description{
This function obtains point-level scores for a match from the sofascore site
}
